+++ examples/flava/native/train.py
@@ -53,8 +53,10 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from torch.utils.tensorboard import SummaryWriter
 from torchmultimodal.models.flava.image_encoder import ImageTransformer
 from torchmultimodal.models.flava.text_encoder import BERTTextEncoder
-from torchmultimodal.models.flava.transformer import FLAVATransformerWithoutEmbeddings
-from torchmultimodal.modules.layers.transformer import TransformerEncoderLayer
+from torchmultimodal.models.flava.transformer import (
+    FLAVATransformerWithoutEmbeddings,
+    TransformerEncoderLayer,
+)
 from torchmultimodal.modules.losses.flava import FLAVAPretrainingLossOutput
