+++ examples/flava/requirements.txt
@@ -7,3 +7,4 @@ omegaconf==2.1.2
 hydra-core==1.1.2
 transformers==4.30.0
 pycocotools==2.0.4
+tensorboard
