+++ tests/models/flava/test_text_encoder.py
@@ -11,10 +11,12 @@ import pytest
 import torch
 from tests.test_utils import assert_expected, set_rng_seed
 from torch import nn
-from torchmultimodal.models.flava.transformer import init_transformer_weights
+from torchmultimodal.models.flava.transformer import (
+    init_transformer_weights,
+    TransformerEncoder,
+)
 from torchmultimodal.modules.encoders.bert_text_encoder import BERTTextEncoder
 from torchmultimodal.modules.layers.text_embedding import BERTTextEmbeddings
-from torchmultimodal.modules.layers.transformer import TransformerEncoder
 
 
 @pytest.fixture(autouse=True)
