+++ torchmultimodal/modules/encoders/bert_text_encoder.py
@@ -8,11 +8,9 @@ from typing import Callable, Optional
 
 import torch
 from torch import nn, Tensor
+from torchmultimodal.models.flava.transformer import TransformerEncoder
 from torchmultimodal.modules.layers.text_embedding import BERTTextEmbeddings
-from torchmultimodal.modules.layers.transformer import (
-    TransformerEncoder,
-    TransformerOutput,
-)
+from torchmultimodal.modules.layers.transformer import TransformerOutput
 from torchmultimodal.utils.attention import get_extended_attention_mask
